Finished command-line interface; added build.xml.
CLI has a default for writing successful matches to STDOUT in a simple
format, as well as a verbose-output mode (written to files) as well as
an interactive mode for viewing results & writing results to files.

Refactored file layout -- core functions are in _core, utils in _util,
and the code for -main & CLI is in modernize.clj. Dropped dependency on
my own file-io library; am using contrib.duck-streams now that it's
encoding-aware.