Added to-java-static-map & a 2.0 stylesheet for replacing text.
to-java-static-map creates a Java class with a static map of words to
their replacements as well as static methods `contains` and `get` for
the standard map operations. to-java-static-map expects to be piped the
output from modernize.main.

text-replace.xsl should be run on a TEI file from the same directory the
Java static map (called Lookup in text-replace) lives.

Note: in modernize.main/run-perms-test, am now using map instead of pmap
-- the latter was eating as much heap space as I would give it (I tried
up to 1GB).